requirements.txt
flask
line-bot-sdk
openai
load_dotenv
langchain
langchain-core
langchain-openai
langchain-community
faiss-cpu